Add precommit gate, format/check targets, warnings-as-errors, and CONTRIBUTING (Phase 4y)
- CMake: unit-test executables moved to EXCLUDE_FROM_ALL; new `tests`
  target builds them, `check` builds + runs ctest
- CMake: `format-tree` / `format-check-tree` targets run clang-format
  over src/, include/promeki/, tests/, demos/, utils/ (-tree suffix
  avoids collision with vendored submodule bare names)
- CMake: `promeki::warnings` interface library applies -Wall -Wextra
  -Wno-unused-parameter -Wno-maybe-uninitialized
  -Wno-missing-field-initializers to all our targets (not thirdparty/);
  PROMEKI_WARNINGS_AS_ERRORS option (default OFF, precommit flips ON)
- scripts/precommit.sh: 5-step gate (configure, format-check-tree,
  full -Werror build, check, doxygen); isolated timestamped build dir;
  non-blocking flock so concurrent invocations exit 3 cleanly
- platform.h: PROMEKI_COMPILER_NVCC macro for CUDA TUs;
  PROMEKI_UNROLL portable loop-unroll hint (Clang/nvcc: _Pragma("unroll"),
  GCC: no-op to suppress -Wunknown-pragmas); system.h migrated from
  naked #pragma unroll
- Warning cleanup (~30 sites): dead helper functions removed,
  sign-compare casts, ctor mem-init order, -Wclobbered around setjmp
  (libjpeg encode path), using Layout::addWidget to un-hide inherited
  overload, pragma guards for intentional strncpy truncation and
  self-move test
- Behavioral fix: AnsiStream::getCursorPosition now returns false
  immediately when the 20-char read loop exits without R terminator,
  preventing parse of truncated garbage; covered by new unit test
- Docs: CONTRIBUTING.md written from scratch; docs/building.md updated
  with tests/format/precommit sections; README.md short note added;
  .gitignore gains .precommit.lock
- devplan/README.md: Phase 4y entry added

+++ src/core/numnameseq.cpp
@@ -29,8 +29,8 @@ NumNameSeq::List NumNameSeq::parseList(StringList &list) {
                         auto &n = i->_name;
                         if (name.isInSequence(n)) {
                                 if ((name.isPadded() && !n.isPadded()) || (name.digits() > n.digits())) n = name;
-                                if (value < i->_head) i->_head = value;
-                                if (value > i->_tail) i->_tail = value;
+                                if (static_cast<size_t>(value) < i->_head) i->_head = static_cast<size_t>(value);
+                                if (static_cast<size_t>(value) > i->_tail) i->_tail = static_cast<size_t>(value);
                                 item = list.remove(item);
                                 found = true;
                                 break;
